+++ services/utils.go
@@ -10,9 +10,6 @@ import (
 	"github.com/manifoldco/promptui"
 
 	"github.com/logiqai/logiqctl/api/v1/query"
-	"github.com/logiqai/logiqctl/cfg"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
 var templates = &promptui.SelectTemplates{
@@ -20,38 +17,6 @@ var templates = &promptui.SelectTemplates{
 	Active:   "\U000000BB {{ .Name | green }} ({{ .Details | red }})",
 	Inactive: "  {{ .Name | cyan }} ({{ .Details | red }})",
 	Selected: "Application {{ .Name | red }} selected",
-}
-
-func handleError(config *cfg.Config, err error) {
-	//fmt.Printf("%v\n",err)
-	errorStatus := status.Convert(err)
-	switch errorStatus.Code() {
-	case codes.Unavailable:
-		fmt.Printf("Could not connect, Are you sure you can connect to [%s]?", config.Cluster)
-		fmt.Println("Please verify your configuration...")
-		fmt.Println("++++++++++++++++++++++")
-		fmt.Print(config.String())
-		fmt.Println("++++++++++++++++++++++")
-		os.Exit(1)
-	case codes.PermissionDenied:
-		fmt.Printf("You are not allowed to do this operations, this may require Administrator privileges\n")
-		os.Exit(1)
-	case codes.Unauthenticated:
-		fmt.Printf("Invalid Token, You may need to Login\n")
-		os.Exit(1)
-	case codes.AlreadyExists:
-		fmt.Printf("Error : %s\n", errorStatus.Message())
-		os.Exit(1)
-	case codes.NotFound:
-		fmt.Printf("Error : %s\n", errorStatus.Message())
-		os.Exit(1)
-	case codes.InvalidArgument:
-		fmt.Printf("Error : %s\n", errorStatus.Message())
-		os.Exit(1)
-	default:
-		fmt.Printf("Error : [%s]\n", errorStatus.Message())
-		break
-	}
 }
 
 const (
@@ -72,7 +37,15 @@ func printSyslogMessage(logMap map[string]interface{}, output string) {
 			logMap["facility_string"],
 			logMap["message"],
 		)
-	} else if output == OUTPUT_RAW {
+	} else if output == OUTPUT_JSON {
+		v, err := json.Marshal(logMap)
+		if err == nil {
+			fmt.Printf("%s\n", v)
+		} else {
+			fmt.Printf("Error marshalling JSON %v", logMap)
+			os.Exit(-1)
+		}
+	} else {
 		fmt.Printf("%s %s %s %s %s %s\n",
 			logMap["timestamp"],
 			logMap["namespace"],
@@ -81,13 +54,8 @@ func printSyslogMessage(logMap map[string]interface{}, output string) {
 			logMap["facility_string"],
 			logMap["message"],
 		)
-	} else if output == OUTPUT_JSON {
-		v, err := json.Marshal(logMap)
-		if err == nil {
-			fmt.Printf("%s\n", v)
-		} else {
-			fmt.Printf("Error marshalling JSON %v", logMap)
-			os.Exit(-1)
+		if utils.GetLineBreak() {
+			fmt.Println()
 		}
 	}
 }
